Move deduplicate function to util.unique
 Also do some minor cosmetic refactoring
