+++ tests/test_gen.py
@@ -192,9 +192,3 @@ class GenTest(unittest.TestCase):
                     u'licenses:\n'
                     u'    -   file: this.LICENSE\n')
         assert expected == in_mem_result
-
-    def test_deduplicate(self):
-        items = ['a', 'b', 'd', 'b', 'c', 'a']
-        expected = ['a', 'b', 'd', 'c']
-        results = gen.deduplicate(items)
-        assert expected == results
